[EXERCISE] 4.15 - 4.19

+++ part4/blog/controllers/blog.js
@@ -1,13 +1,24 @@
 const blogRouter = require('express').Router()
-const Blog = require('../models/blog')
+const Blog = require('../models/blog')  
+const User = require('../models/user') 
 
 blogRouter.get('/', async (req, res) => {
-    const blogs = await Blog.find({})
-    res.status(200).json(blogs)
+    try {
+        const blogs = await Blog
+            .find({})
+            .populate('author', { username: 1, name: 1 }) 
+
+        res.status(200).json(blogs)
+    } catch (error) {
+        res.status(400).json({ error: error.message })
+    }
 })
 
+
 blogRouter.get('/:id', async (req, res) => {
+
     const blog = await Blog.findById(req.params.id)
+
     if (blog) {
         res.status(200).json(blog)
     } else {
@@ -16,39 +27,79 @@ blogRouter.get('/:id', async (req, res) => {
 })
 
 blogRouter.post('/', async (req, res) => {
-    const {
-        title,
-        author,
-        url,
-        likes
-    } = req.body
+    const { title, author, url, likes } = req.body
 
-    const blog = new Blog({
-        title,
-        author,
-        url,
-        likes
-    })
-    const saveBlog = await blog.save()
-    res.status(201).json(saveBlog)
+    if (!title || !author || !url) {
+        return res.status(400).json({ error: 'Title, author, and URL are required' })
+    }
+
+    try {
+        
+        const blog = new Blog({
+            title,
+            author,
+            url,
+            likes: likes || 0, // Si no se pasa "likes", se asigna 0
+        })
+
+        // Guardar el blog en la base de datos
+        const savedBlog = await blog.save()
+
+        // Agregar el blog al campo 'blogs' del usuario
+        const user = await User.findById(author)
+        user.blogs = user.blogs.concat(savedBlog._id)  // Agregar el ID del nuevo blog al array 'blogs'
+        await user.save()
+
+        res.status(201).json(savedBlog)
+    } catch (error) {
+       
+        res.status(400).json({ error: error.message })
+    }
 })
+
 
 blogRouter.put('/:id', async (req, res) => {
     const { title, author, url, likes } = req.body
-    const updatedBlog = await Blog.findByIdAndUpdate(
-        req.params.id,
-        { title, author, url, likes },
-        { new: true, runValidators: true } 
-    )
-    if (!updatedBlog) {
-        return res.status(404).json({ error: "Blog NOT found" })
+
+    if (!title || !author || !url) {
+        return res.status(400).json({ error: 'Title, author, and URL are required' })
     }
-    res.status(200).json(updatedBlog)
+
+    try {
+        const updatedBlog = await Blog.findByIdAndUpdate(
+            req.params.id,
+            {
+                title,
+                author,
+                url,
+                likes: likes || 0, // Si no se pasa "likes", se asigna 0
+            },
+            {
+                new: true,
+                runValidators: true,
+            }
+        )
+
+        if (!updatedBlog) {
+            return res.status(404).json({ error: 'Blog not found' })
+        }
+
+        res.status(200).json(updatedBlog)
+    } catch (error) {
+        res.status(400).json({ error: error.message })
+    }
 })
 
 blogRouter.delete('/:id', async (req, res) => {
+    const blog = await Blog.findById(req.params.id)
+
+    if (!blog) {
+        return res.status(404).json({ error: 'Blog not found' })
+    }
+
     await Blog.findByIdAndDelete(req.params.id)
     res.status(204).end()
 })
+
 
 module.exports = blogRouter
